+++ scripts/check-gmp.gmp.py
@@ -131,15 +131,13 @@ class InstanceManager:
 
         except PermissionError:
             parser.error(
-                "The selected temporary database file {} or the parent dir has not the correct permissions.".format(
-                    self.db
-                )
+                "The selected temporary database file {} or the parent dir has"
+                " not the correct permissions.".format(self.db)
             )
 
     @staticmethod
     def _to_sql_bool(pending):
-        """ Replace True/False with 1/0.
-        """
+        """Replace True/False with 1/0."""
         return '1' if pending else '0'
 
     def connect_db(self):
@@ -149,15 +147,14 @@ class InstanceManager:
         """
         try:
             logger.debug("connect db: %s", self.db)
-            self.con_db = sqlite3.connect(self.db)
+            self.con_db = sqlite3.connect(str(self.db))
             self.cursor = self.con_db.cursor()
             logger.debug(sqlite3.sqlite_version)
         except Exception as e:  # pylint: disable=broad-except
             logger.debug(e)
 
     def close_db(self):
-        """Close database
-        """
+        """Close database"""
         self.con_db.close()
 
     def set_host(self, host):
@@ -259,8 +256,7 @@ class InstanceManager:
         self.con_db.commit()
 
     def delete_report(self):
-        """Delete report from database
-        """
+        """Delete report from database"""
         self.cursor.execute("DELETE FROM Report WHERE host=?", (self.host,))
 
         # Save the changes
@@ -276,7 +272,7 @@ class InstanceManager:
         self.cursor.execute("DELETE FROM Report WHERE host=?", (ip,))
         self.con_db.isolation_level = None
         self.cursor.execute("VACUUM")
-        self.con_db.isolation_level = ''  # see: https://github.com/CxAalto/gtfspy/commit/8d05c3c94a6d4ca3ed675d88af93def7d5053bfe
+        self.con_db.isolation_level = ''  # see: https://github.com/CxAalto/gtfspy/commit/8d05c3c94a6d4ca3ed675d88af93def7d5053bfe # pylint: disable=line-too-long
         # Save the changes
         self.con_db.commit()
 
@@ -656,15 +652,15 @@ def status(gmp, im, script_args):
                 full_report = gmp.get_report(
                     report_id=last_report_id,
                     filter="sort-reverse=id result_hosts_only=1 "
-                    "min_cvss_base= min_qod= levels=hmlgd autofp=%s "
-                    "notes=0 apply_overrides=%s overrides=%s first=1 rows=-1 "
-                    "delta_states=cgns host=%s"
-                    % (
+                    "min_cvss_base= min_qod= levels=hmlgd autofp={} "
+                    "notes=0 apply_overrides={} overrides={} first=1 rows=-1 "
+                    "delta_states=cgns host={}".format(
                         script_args.autofp,
                         int(script_args.overrides),
                         int(script_args.apply_overrides),
                         host,
                     ),
+                    details=True,
                 )
 
                 im.add_report(last_scan_end, params_used, full_report)
@@ -1035,9 +1031,7 @@ ZERO = timedelta(0)
 
 
 class Utc(tzinfo):
-    """UTC Timezone
-
-    """
+    """UTC Timezone"""
 
     def utcoffset(self, dt):
         return ZERO
@@ -1056,9 +1050,7 @@ UTC = Utc()
 
 
 class FixedOffset(tzinfo):
-    """Fixed offset in hours and minutes from UTC
-
-    """
+    """Fixed offset in hours and minutes from UTC"""
 
     def __init__(self, offset_hours, offset_minutes, name):
         self.__offset_hours = offset_hours  # Keep for later __getinitargs__
@@ -1117,9 +1109,7 @@ def to_int(
 
 
 def parse_timezone(matches, default_timezone=UTC):
-    """Parses ISO 8601 time zone specs into tzinfo offsets
-
-    """
+    """Parses ISO 8601 time zone specs into tzinfo offsets"""
 
     if matches["timezone"] == "Z":
         return UTC
@@ -1133,8 +1123,8 @@ def parse_timezone(matches, default_timezone=UTC):
     minutes = to_int(matches, "tz_minute", default_to_zero=True)
     description = "%s%02d:%02d" % (sign, hours, minutes)
     if sign == "-":
-        hours = -hours
-        minutes = -minutes
+        hours = -1 * hours
+        minutes = -1 * minutes
     return FixedOffset(hours, minutes, description)
 
 
@@ -1194,7 +1184,7 @@ def parse_date(datestring, default_timezone=UTC):
             tzinfo=tz,
         )
     except Exception as e:
-        raise ParseError(e)
+        raise ParseError(e) from None
 
 
 def main(gmp, args):